refactor : 컴포넌트 계층화 하기
